+++ mycodo/scripts/upgrade_commands.sh
@@ -25,7 +25,7 @@ INFLUXDB2_VERSION="2.6.1"
 VIRTUALENV_VERSION="20.17.1"
 
 # Required apt packages
-APT_PKGS="gawk gcc g++ git jq libatlas-base-dev libffi-dev libi2c-dev logrotate moreutils netcat nginx python3 python3-pip python3-dev python3-setuptools rng-tools sqlite3 unzip wget"
+APT_PKGS="gawk gcc g++ git jq libatlas-base-dev libffi-dev libi2c-dev logrotate moreutils netcat-openbsd nginx python3 python3-pip python3-dev python3-setuptools rng-tools sqlite3 unzip wget"
 
 PYTHON_BINARY_SYS_LOC="$(python3 -c "import os; print(os.environ['_'])")"
 
@@ -586,9 +586,9 @@ case "${1:-''}" in
     ;;
     'update-permissions')
         printf "\n#### Setting permissions\n"
-        chown -LR mycodo.mycodo "${MYCODO_PATH}"
-        chown -R mycodo.mycodo /var/log/mycodo
-        chown -R mycodo.mycodo /var/Mycodo-backups
+        chown -LR mycodo:mycodo "${MYCODO_PATH}"
+        chown -R mycodo:mycodo /var/log/mycodo
+        chown -R mycodo:mycodo /var/Mycodo-backups
 
         find "${MYCODO_PATH}" -type d -exec chmod u+wx,g+wx {} +
         find "${MYCODO_PATH}" -type f -exec chmod u+w,g+w,o+r {} +