+++ mycodo/scripts/upgrade_commands.sh
@@ -602,7 +602,7 @@ case "${1:-''}" in
         if [[ ! -d ${MYCODO_PATH}/env ]]; then
             printf "\n## Error: Virtualenv doesn't exist. Create with %s setup-virtualenv\n" "${0}"
         else
-            "${MYCODO_PATH}"/env/bin/python -m pip install --break-system-packages --upgrade pip
+            "${MYCODO_PATH}"/env/bin/python -m pip install --upgrade pip
         fi
     ;;
     'update-pip3-packages')
@@ -611,8 +611,8 @@ case "${1:-''}" in
         if [[ ! -d ${MYCODO_PATH}/env ]]; then
             printf "\n## Error: Virtualenv doesn't exist. Create with %s setup-virtualenv\n" "${0}"
         else
-            "${MYCODO_PATH}"/env/bin/python -m pip install --break-system-packages --upgrade -r "${MYCODO_PATH}"/install/requirements.txt
-            "${MYCODO_PATH}"/env/bin/python -m pip install --break-system-packages --upgrade -r "${MYCODO_PATH}"/install/requirements-testing.txt
+            "${MYCODO_PATH}"/env/bin/python -m pip install --upgrade -r "${MYCODO_PATH}"/install/requirements.txt
+            "${MYCODO_PATH}"/env/bin/python -m pip install --upgrade -r "${MYCODO_PATH}"/install/requirements-testing.txt
         fi
     ;;
     'pip-clear-cache')
@@ -752,7 +752,7 @@ case "${1:-''}" in
         if [[ ! -d ${MYCODO_PATH}/env ]]; then
             printf "\n## Error: Virtualenv doesn't exist. Create with %s setup-virtualenv\n" "${0}"
         else
-            "${MYCODO_PATH}"/env/bin/python -m pip install --break-system-packages --upgrade pip
+            "${MYCODO_PATH}"/env/bin/python -m pip install --upgrade pip
         fi
     ;;
     'docker-update-pip-packages')
@@ -761,7 +761,7 @@ case "${1:-''}" in
         if [[ ! -d ${MYCODO_PATH}/env ]]; then
             printf "\n## Error: Virtualenv doesn't exist. Create with %s setup-virtualenv\n" "${0}"
         else
-            "${MYCODO_PATH}"/env/bin/python -m pip install --break-system-packages --no-cache-dir -r "${MYCODO_PATH}"/install/requirements.txt
+            "${MYCODO_PATH}"/env/bin/python -m pip install --no-cache-dir -r "${MYCODO_PATH}"/install/requirements.txt
         fi
     ;;
     'install-docker')